ci: remove auto-gen docs action. Prep for new approach
This represents a rethink of approach in the api docs creation. The primary reason for this is that
auto-generating the documentation on every push will cause the documentation to be out-of-sync with
the published codebase, which could lead to confusion. Ultimately, we'll need a proper release
process so that the documentation is tied to a release version, but that's likely a step too far at
this stage.

This commit removes the github action and re-enables the links from the generate/publish-all
scripts. The publish script for the docs will follow