Pull from smart core os.GitHub.io
The aim of this change is to trigger `smart-core-os/smart-core-os.github.io` to build the documentation for SmartCore itself whenever a change is made to the `sc-api`.

Before, `sc-api` would execute a script to generate this documentation and push the documentation changes to `smart-core-os.github.io`.

Now, `smart-core-os.github.io` pulls the protobuf changes and builds the documentation in whatever format it decides is best for readers.