protobuf: git tag v1.0.0-beta.55 uses protoc v5.29.3 (#28)
The purpose of this PR is to create git tag v1.0.0-beta.55 using the same protoc version (v5.29.3) in all implementations.

The commit also includes a tag for the protobuf directory, such that proto.mod usage is able to fetch a consistent view of protobuf vs the actual implementations.

Previously, hail.proto was missing definitions for DescribeHail even if the file had already been committed with these changes to this repository without the necessary git tag.

* chore: publish v1.0.0-beta.55

* protobuf: tag all protobuf implementations to version 1.0.0-beta.55